+++ app/Http/Controllers/API/V1/LatestPosts.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\API\V1;
 
-use App\Http\Resources\PostResource;
+use App\Http\Resources\API\V1\PostResource;
 use App\Models\Post;
 use Illuminate\Http\Request;
