+++ app/Http/Resources/API/V1/TagResource.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Http\Resources;
+namespace App\Http\Resources\API\V1;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class UserResource extends JsonResource
+class TagResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
@@ -15,10 +15,9 @@ class UserResource extends JsonResource
     public function toArray(Request $request): array
     {
         return [
-            'username' => $this->username,
             'name' => $this->name,
-            'avatar' => $this->avatar,
-            'role' => $this->role->name
+            'slug' => $this->slug,
+            'color' => $this->color
         ];
     }
 }
