+++ app/Http/Resources/API/V1/UserResource.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Http\Resources;
+namespace App\Http\Resources\API\V1;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class TagResource extends JsonResource
+class UserResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
@@ -15,9 +15,10 @@ class TagResource extends JsonResource
     public function toArray(Request $request): array
     {
         return [
+            'username' => $this->username,
             'name' => $this->name,
-            'slug' => $this->slug,
-            'color' => $this->color
+            'avatar' => $this->avatar,
+            'role' => $this->role->name
         ];
     }
 }
